requirements.txt
# Python dependencies for SAB SMS template discovery pipeline
numpy>=1.24.0
scikit-learn>=1.3.0
pyyaml>=6.0
